pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "heartlib"
version = "0.1.0"
description = "A Python Library."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "HeartMuLa Team", email = "heartmula.ai@gmail.com"}
]
dependencies = [
    "numpy==2.0.2",
    # torch, torchvision, torchaudio are installed separately with CUDA support by setup_gui.bat
    # DO NOT add them here as it will install CPU-only versions
    "torchtune==0.4.0",
    "torchao==0.9.0",
    "tqdm==4.67.1",
    "traitlets==5.7.1",
    "traittypes==0.2.3",
    "transformers==4.57.0",
    "tokenizers==0.22.1",
    "ipykernel==6.17.1",
    "einops==0.8.1",
    "accelerate==1.12.0",
    "bitsandbytes==0.49.0",
    "vector-quantize-pytorch==1.27.15",
    "modelscope==1.33.0",
    "soundfile"
]
urls = { "homepage" = "https://heartmula.github.io/" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

